Fix tracking not working after internal restart while FPFC is enabled
