+++ SiraUtil/manifest.json
@@ -3,7 +3,7 @@
   "id": "SiraUtil",
   "name": "SiraUtil",
   "author": "Auros",
-  "version": "3.1.8",
+  "version": "3.1.9",
   "icon": "SiraUtil.Resources.logo.png",
   "description": "A powerful utility mod which provides more tools to Beat Saber modders.",
   "gameVersion": "1.35.0",
